app/requirements.txt
google-search-results
langchain
openai
python-dotenv
superset-api-client
wikipedia
